+++ lib/services/doorman.dart
@@ -43,6 +43,7 @@ class DoormanServiceWithToken {
     }
 
     return CallOptions(
+      timeout: const Duration(seconds: 5),
       metadata: {
         "Authorization": "Bearer $token",
       },