+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unrealised]
+
+### Changed
+
+- Change event link icon to chat bubble if applicable
+- Fix contact color on event page
+
 ## [2.0.0] - 2023-11-17
 
 ### Changed
@@ -18,13 +25,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [1.2.3] - 2023-11-04
 
-### Fix
+### Fixed
 
 - App versioning
 
 ## [1.2.2] - 2023-11-02
 
-### Fix
+### Fixed
 
 - Increase timeout for doorman connection